LINE: remove shared group mentions helper
(cherry picked from commit 1aae93b1fa67f5133ff35d6f999a848f767f1225)
