+++ extensions/line/src/group-policy.ts
@@ -1,0 +1,22 @@
+import { resolveChannelGroupRequireMention } from "openclaw/plugin-sdk/channel-policy";
+import { resolveExactLineGroupConfigKey, type OpenClawConfig } from "openclaw/plugin-sdk/line-core";
+
+type LineGroupContext = {
+  cfg: OpenClawConfig;
+  accountId?: string | null;
+  groupId?: string | null;
+};
+
+export function resolveLineGroupRequireMention(params: LineGroupContext): boolean {
+  const exactGroupId = resolveExactLineGroupConfigKey({
+    cfg: params.cfg,
+    accountId: params.accountId,
+    groupId: params.groupId,
+  });
+  return resolveChannelGroupRequireMention({
+    cfg: params.cfg,
+    channel: "line",
+    groupId: exactGroupId ?? params.groupId,
+    accountId: params.accountId,
+  });
+}
