--- src/channels/plugins/group-mentions.test.ts
@@ -1,201 +1,0 @@
-import { describe, expect, it } from "vitest";
-import {
-  resolveBlueBubblesGroupRequireMention,
-  resolveBlueBubblesGroupToolPolicy,
-  resolveLineGroupRequireMention,
-  resolveLineGroupToolPolicy,
-  resolveSlackGroupRequireMention,
-  resolveSlackGroupToolPolicy,
-  resolveTelegramGroupRequireMention,
-  resolveTelegramGroupToolPolicy,
-} from "./group-mentions.js";
-
-const cfg = {
-  channels: {
-    slack: {
-      botToken: "xoxb-test",
-      appToken: "xapp-test",
-      channels: {
-        alerts: {
-          requireMention: false,
-          tools: { allow: ["message.send"] },
-          toolsBySender: {
-            "id:user:alice": { allow: ["sessions.list"] },
-          },
-        },
-        "*": {
-          requireMention: true,
-          tools: { deny: ["exec"] },
-        },
-      },
-    },
-  },
-  // oxlint-disable-next-line typescript/no-explicit-any
-} as any;
-
-describe("group mentions (slack)", () => {
-  it("uses matched channel requireMention and wildcard fallback", () => {
-    expect(resolveSlackGroupRequireMention({ cfg, groupChannel: "#alerts" })).toBe(false);
-    expect(resolveSlackGroupRequireMention({ cfg, groupChannel: "#missing" })).toBe(true);
-  });
-
-  it("resolves sender override, then channel tools, then wildcard tools", () => {
-    const senderOverride = resolveSlackGroupToolPolicy({
-      cfg,
-      groupChannel: "#alerts",
-      senderId: "user:alice",
-    });
-    expect(senderOverride).toEqual({ allow: ["sessions.list"] });
-
-    const channelTools = resolveSlackGroupToolPolicy({
-      cfg,
-      groupChannel: "#alerts",
-      senderId: "user:bob",
-    });
-    expect(channelTools).toEqual({ allow: ["message.send"] });
-
-    const wildcardTools = resolveSlackGroupToolPolicy({
-      cfg,
-      groupChannel: "#missing",
-      senderId: "user:bob",
-    });
-    expect(wildcardTools).toEqual({ deny: ["exec"] });
-  });
-});
-
-describe("group mentions (telegram)", () => {
-  it("resolves topic-level requireMention and chat-level tools for topic ids", () => {
-    const telegramCfg = {
-      channels: {
-        telegram: {
-          botToken: "telegram-test",
-          groups: {
-            "-1001": {
-              requireMention: true,
-              tools: { allow: ["message.send"] },
-              topics: {
-                "77": {
-                  requireMention: false,
-                },
-              },
-            },
-            "*": {
-              requireMention: true,
-            },
-          },
-        },
-      },
-      // oxlint-disable-next-line typescript/no-explicit-any
-    } as any;
-    expect(
-      resolveTelegramGroupRequireMention({ cfg: telegramCfg, groupId: "-1001:topic:77" }),
-    ).toBe(false);
-    expect(resolveTelegramGroupToolPolicy({ cfg: telegramCfg, groupId: "-1001:topic:77" })).toEqual(
-      {
-        allow: ["message.send"],
-      },
-    );
-  });
-});
-
-describe("group mentions (bluebubbles)", () => {
-  it("uses generic channel group policy helpers", () => {
-    const blueBubblesCfg = {
-      channels: {
-        bluebubbles: {
-          groups: {
-            "chat:primary": {
-              requireMention: false,
-              tools: { deny: ["exec"] },
-            },
-            "*": {
-              requireMention: true,
-              tools: { allow: ["message.send"] },
-            },
-          },
-        },
-      },
-      // oxlint-disable-next-line typescript/no-explicit-any
-    } as any;
-
-    expect(
-      resolveBlueBubblesGroupRequireMention({ cfg: blueBubblesCfg, groupId: "chat:primary" }),
-    ).toBe(false);
-    expect(
-      resolveBlueBubblesGroupRequireMention({ cfg: blueBubblesCfg, groupId: "chat:other" }),
-    ).toBe(true);
-    expect(
-      resolveBlueBubblesGroupToolPolicy({ cfg: blueBubblesCfg, groupId: "chat:primary" }),
-    ).toEqual({ deny: ["exec"] });
-    expect(
-      resolveBlueBubblesGroupToolPolicy({ cfg: blueBubblesCfg, groupId: "chat:other" }),
-    ).toEqual({
-      allow: ["message.send"],
-    });
-  });
-});
-
-describe("group mentions (line)", () => {
-  it("matches raw and prefixed LINE group keys for requireMention and tools", () => {
-    const lineCfg = {
-      channels: {
-        line: {
-          groups: {
-            "room:r123": {
-              requireMention: false,
-              tools: { allow: ["message.send"] },
-            },
-            "group:g123": {
-              requireMention: false,
-              tools: { deny: ["exec"] },
-            },
-            "*": {
-              requireMention: true,
-            },
-          },
-        },
-      },
-      // oxlint-disable-next-line typescript/no-explicit-any
-    } as any;
-
-    expect(resolveLineGroupRequireMention({ cfg: lineCfg, groupId: "r123" })).toBe(false);
-    expect(resolveLineGroupRequireMention({ cfg: lineCfg, groupId: "room:r123" })).toBe(false);
-    expect(resolveLineGroupRequireMention({ cfg: lineCfg, groupId: "g123" })).toBe(false);
-    expect(resolveLineGroupRequireMention({ cfg: lineCfg, groupId: "group:g123" })).toBe(false);
-    expect(resolveLineGroupRequireMention({ cfg: lineCfg, groupId: "other" })).toBe(true);
-    expect(resolveLineGroupToolPolicy({ cfg: lineCfg, groupId: "r123" })).toEqual({
-      allow: ["message.send"],
-    });
-    expect(resolveLineGroupToolPolicy({ cfg: lineCfg, groupId: "g123" })).toEqual({
-      deny: ["exec"],
-    });
-  });
-
-  it("uses account-scoped prefixed LINE group config for requireMention", () => {
-    const lineCfg = {
-      channels: {
-        line: {
-          groups: {
-            "*": {
-              requireMention: true,
-            },
-          },
-          accounts: {
-            work: {
-              groups: {
-                "group:g123": {
-                  requireMention: false,
-                },
-              },
-            },
-          },
-        },
-      },
-      // oxlint-disable-next-line typescript/no-explicit-any
-    } as any;
-
-    expect(
-      resolveLineGroupRequireMention({ cfg: lineCfg, groupId: "g123", accountId: "work" }),
-    ).toBe(false);
-  });
-});
